move Navigation component from references to core
+ rename 'Submit' to 'Finish'
